Split dependencies into their owning functions' files
