+++ src/converters/comments/convertFileComments.ts
@@ -1,11 +1,18 @@
 import { FileSystem } from "../../adapters/fileSystem";
+import { fsFileSystem } from "../../adapters/fsFileSystem";
 import { RuleConverter } from "../lintConfigs/rules/ruleConverter";
+import { ruleConverters } from "../lintConfigs/rules/ruleConverters";
 import { parseFileComments } from "./parseFileComments";
 import { replaceFileComments } from "./replaceFileComments";
 
 export type ConvertFileCommentsDependencies = {
     converters: Map<string, RuleConverter>;
     fileSystem: Pick<FileSystem, "readFile" | "writeFile">;
+};
+
+export const convertFileCommentsDependencies: ConvertFileCommentsDependencies = {
+    converters: ruleConverters,
+    fileSystem: fsFileSystem,
 };
 
 export const convertFileComments = async (
