+++ src/converters/lintConfigs/reporting/packages/choosePackageManager.ts
@@ -1,8 +1,13 @@
 import { FileSystem } from "../../../../adapters/fileSystem";
+import { fsFileSystem } from "../../../../adapters/fsFileSystem";
 import { preferredLockfiles, PackageManager } from "./packageManagers";
 
 export type ChoosePackageManagerDependencies = {
     fileSystem: Pick<FileSystem, "fileExists">;
+};
+
+export const choosePackageManagerDependencies: ChoosePackageManagerDependencies = {
+    fileSystem: fsFileSystem,
 };
 
 export const choosePackageManager = async (dependencies: ChoosePackageManagerDependencies) => {
